+++ UiLepra/Include/Win32/UiWin32Input.h
@@ -130,8 +130,8 @@ public:
 	virtual void ShowCursor();
 	virtual void HideCursor();
 
-	virtual double GetCursorX();
-	virtual double GetCursorY();
+	virtual float GetCursorX();
+	virtual float GetCursorY();
 
 	bool IsInitialized();
 
@@ -162,8 +162,8 @@ private:
 	int mScreenHeight;
 
 	// Mouse related stuff.
-	double mCursorX;
-	double mCursorY;
+	float mCursorX;
+	float mCursorY;
 
 	// Default devices.
 	InputDevice* mKeyboard;
